requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Admin==1.6.1
Flask-Migrate==4.0.4
mysql-connector-python==8.0.32
python-dotenv==1.0.0
requests==2.28.2
openai==0.27.2
Flask-WTF==1.1.1
Pillow==11.0.0
Werkzeug==2.2.3
gunicorn==20.1.0
PyMySQL==1.0.3
reportlab==4.0.4